Call update() in load()
`this.els.link.href = this.makeHash();` in main() is now redundant,
and it was also incorrect since it was not being executed on hashchange,
causing the permalink to be incorrect after a hash change.

A doState parameter (default: true) is added to update() because the
history state does not always need to be updated when the document is
loaded.