Fix re-opening the same file in Blink/WebKit
The `<input type="file">`'s list of selected files was not being
cleared after opening the file.  In Firefox, reopening a file that is
already selected does trigger a `change` event, but not in Blink or
WebKit.  This fixes that by setting the element's `value` to the empty
string immediately after reading the file's text.